+++ libraries/global_config_options.rb
@@ -6,8 +6,9 @@ module CaOpenldap
     def initialize(path="/etc/openldap/slapd.d/cn=config.ldif")
       @path = path
 
-      content = File.read(path).split(/\n/)
-      olc_lines = content.grep /^olc/
+      content = read_lines_from_file @path
+
+      olc_lines = content.grep(/^olc/)
       @options = olc_lines.inject({}) do |hash, line|
         insert_option_to_hash(hash, line)
       end
@@ -26,7 +27,12 @@ module CaOpenldap
       current = {}
 
       File.open(updated, "w") do |output|
-        File.read(@path).split(/\n/).each do |line|
+
+        content = read_lines_from_file @path
+        content.each do |line|
+
+          # Remove CRC comment
+          next if line.start_with? '# CRC'
 
           # Output options just before the structuralObjectClass line.
           if line.start_with? 'structuralObjectClass'
@@ -45,7 +51,10 @@ module CaOpenldap
         end
       end
 
+      # Update the file content while preserving its UID/GID
+      old_file_stats = File.stat(@path)
       File.rename(updated, @path)
+      FileUtils.chown(old_file_stats.uid, old_file_stats.gid, @path)
 
       # return true if changes detected.
       @options != current
@@ -58,9 +67,22 @@ module CaOpenldap
     private 
 
     def insert_option_to_hash(hash, line)
-      k, v = line.split(':')
+      k, v = line.split(':', 2)
       hash[k.strip] = v.strip
       hash
+    end
+
+    # Read lines from slapd config file, joining
+    # multi-line definitions as single-ligne definitions.
+    # @param [Array<String>] List of lines found in config file.
+    def read_lines_from_file(file_path)
+      full_content = File.read(file_path)
+
+      # join multi-lines
+      full_content.gsub!(/\n /, "")
+
+      # Split lines
+      full_content.split(/\n/)
     end
 
   end