+++ recipes/client.rb
@@ -17,8 +17,6 @@
 # limitations under the License.
 #
 
-include_recipe 'ca_openldap::default'
-
 class Chef::Recipe
   include CAOpenldap
 end
@@ -27,7 +25,7 @@ package "openldap-clients" do
   action :upgrade
 end
 
-ldap_conf = case node[:platform_family]
+ldap_conf = case node['platform_family']
 when "rhel"
   "/etc/openldap/ldap.conf"
 else